week 3 adjustments

+++ dbt-greenery/macros/event_types.sql
@@ -1,5 +1,5 @@
 ​​{% macro event_types(event_type) %}
 
-   SUM(CASE WHEN event_type = '{{ event_type }}' THEN 1 ELSE 0 END)    AS  {{ event_type }}
+   SUM(CASE WHEN event_type = '{{ event_type }}' THEN 1 ELSE 0 END)::numeric  AS  {{ event_type }}
 
 {% endmacro %}
